Enhance default action mechanism
There is now a default default action for the current dir.  This is
triggered by the '.' command.  Basically, it compiles main.c, runs it if
successful, and opens errors in Vim quicklist if fails.

You can override the default default dir action for any dir by supplying
a default.sh script in that dir.  If such a script is present, it will
be run when you issue the '.' command.  It does not get sourced--it is
run as a separate process.